# test-constructor/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend part of test constructor",
        "title": "test constructor",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/manager/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать нового организатора",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать организатора",
                "parameters": [
                    {
                        "description": "Manager object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateManagerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/admin.CreateManagerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/intern/tests/{link}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание попытки по ссылке конфигурации мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intern"
                ],
                "summary": "Пройти тест через конфигурацию мероприятия",
                "parameters": [
                    {
                        "description": "Start attempt object",
                        "name": "answers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/intern.StartAttemptRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Ссылка конфигурации теста",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/intern.StartAttemptResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список мероприятий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Получить мероприятия",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/manager.Event"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Создать настройку мероприятия",
                "parameters": [
                    {
                        "description": "EventCfg object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/manager.EventCfgInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/manager/events/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Обновить настройку мероприятия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "EventCfg object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/manager.EventCfgInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.CreateManagerRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "admin.CreateManagerResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "intern.PublicMatching": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "right": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "intern.PublicOptions": {
            "type": "object",
            "properties": {
                "case_sensitive": {
                    "type": "boolean"
                },
                "choice": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "matching": {
                    "$ref": "#/definitions/intern.PublicMatching"
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "intern.QuestionInfo": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/intern.PublicOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.QType"
                }
            }
        },
        "intern.StartAttemptRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "integer"
                }
            }
        },
        "intern.StartAttemptResponse": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "integer"
                },
                "config_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/intern.QuestionInfo"
                    }
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                },
                "time_limit": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "manager.Event": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "manager.EventCfgInfo": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "extra_threshold": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.ExtraThresholdInfo"
                    }
                },
                "fail_text": {
                    "type": "string"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "success_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
...
        "manager.ExtraThresholdInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
...
        "models.QType": {
            "type": "string",
            "enum": [
                "single_choice",
                "multiple_choice",
                "text_input",
                "matching",
                "correct_order"
            ],
            "x-enum-varnames": [
                "SingleChoice",
                "MultipleChoice",
                "TextInput",
                "Matching",
                "CorrectOrder"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}